RELENG-234: restore './configure' to working order
This is now based on targets.json which has all the necessary
repo information.

Also add '-r' to mean "regenerate the config.mk file without
mucking with repos in builds/ or bits in bits/". It *used* to
mean some "reconfigure" middle ground that I don't think was
being used anywhere.