MON-29 expose amon-agent to all zones (MG appliance build 'tarballs')
Update the MG "appliance" builds to require a "tarballs" key in
targets.json that more explicitly lists which tarballs to crack
into the appliance's "foo_dataset" target. Before this, this
was implied by the "deps" array. This functionality will be needed
for adding the amon-agent-*.tgz tarball to the appliance zones.
The amon-agent tarball structure and target location differ from
the fs tarballs.