MON-266: '/opt/amon-agent: No such file or directory' warning in amon zone setup
Also... good to have the agent running in the amon zone to monitor some
non-amon-master direct things. The worry *was* a possible infinite loop
from an amon probe on the amon-master log that was from the code path
handling the event. Just don't set a probe like that an we're fine.