AGENT-732: config-agent handling agent configuration in the global zone
First, agentsshar will need to preload bits from the SAPI build (where
config-agent is built currently). Safe to put that in #master now
because it just means preloading a bit more, even if those bits aren't
used in the build.

Second, use BRANCH as the arg to "mk-agent-shar" because otherwise,
if the AGENTS_INSTALLER_BRANCH is the same as TRY_BRANCH, then
the fallback BRANCH would be removed as a fallback for the other
agent packages.